test-requirements.txt
mock
pytest==3.0.6
pytest-flask==0.8.1
PyYAML
swagger_spec_validator
-e git+https://github.com/uc-cdis/cdisutils-test.git@0.2.2#egg=cdisutilstest
-e git+https://github.com/uc-cdis/indexclient.git@1.5.8#egg=indexclient
